Handle `If` in `BCodeBodyBuilder.genCond()`.
This is more for completeness of `genCond()`. It can now push down
the jumps inside all the control structures for which it makes
sense to do so. It typically applies when the condition of an `if`
or `while` is an `if..else` expression.

It is however not always a win. If the result of the nested `if`s
must be converted to a boolean anyway, doing this will result in
duplication of the conversion to boolean.

A jardiff on the bootstrapped `scala3-compiler.jar` shows 191
additions and 287 deletions, suggesting that, on average, this is
good to have.