Don't simplify lubs when unpickling Ifs
I was not able to find out why that's the case, but the original test
case with the import in the function body requires that if's are _not_
simplified by Unpickler. The interaction with lubs and unsafenulls seems
to be very tricky.

Anyway, we now have two versions of all test cases, with local as well as
globak unsafe nulls.