Remove syntactic deviation for ascriptions in parentheses
Previously the parser accepted `(x: A => B)` as an expression, but
only if there were parentheses around it. This was necessary so that
we could parse a potential formal parameter of a closure `(x: A => B) => ...`
as an expression and then convert to a type. But it is not backed
by the grammar. The grammar requires `x: (A => B)` whether in
parentheses or not.

We now follow the grammar for source versions from 3.2 on.