Don't allow single apply's to be used as constructor proxies
The previous commits allowed them but that caused runtime errors in Specs2 since
`apply` was bound to something different from before.

Since this is a weird corner case it's better not to allow it. As long as we do
not crash we are good.

The error message we get now is:

    constructor proxy `apply` cannot be used as a value

which is also a bit strange, but since this is a truly weird corner case it's not worth it
to finetune this.