Fix loop in parser
I noted a case in Metals where the compiler would keep running at 100+ % until the process was killed.

Using jstack I tracked it down to an infinite `skip` caused by a `syntaxError` in `pattern3`. In fact,
the syntaxError should not skip at this point since the offending expression was already fully parsed.
I fixed this in this commit.

The parser was invoked from the `signatureHelp` method. It seems it parsed something that was not
syntactically correct (specifically, a postfix `*` appeared in a pattern where none was allowed).

`