Better infos for dependent class parameter references
We sometimes create a dependent parameter reference p.X where
`p` is a path with a type that has a wildcard argument, e.g. `P[X >: L <: H].`
So far the denotation of `p.X` had as info the bounds with which `X` was
declared in `P`. Now it gets the actual parameter bounds instead.

Fixes #15652

#15652 started failing when tuple unpackings started to use `val`s instead of `def`s
in #14816. That caused dependent class parameter references to be created and uncovered
the problem.