Scala.js: Handle `@js.native private val`s.
In dotc, `private val`s do not generate a getter method, so always
relying on the getter to generate the `JSNativeMemberDef` and
reading it is not correct.

We now generate `JSNativeMemberDef`s both for `ValDef`s and for
`DefDef`s that are not Accessors. For reading, we support both
`Select`s of fields and `Apply`s of methods.