Fix virtual line numbers in assert-stack
When emitting the bytecode, line numbers are emitted
for trees. These line numbers are used in call stack
frames. Trees inlined from other files get virtual
numbers.

It appears that, when inlining `call(x)`, the line
number for `call` is emitted before that for `x`.
So, if `x` is coming from the current file, its
line number will be real and will override that
of `call`.