Fix initializing parents of AnonClasses
Previously creating ClassDefs of AnonClasses was only possible if the
parent class had a constructor with an empty parameter list. Otherwise,
the compiler would crash because of calling `asTerm` on a `NoSymbol`.
This PR adds logic that handles calling parent constructors with default
or repeated parameters. And also fails more gracefully if no instance can
be created.

fixes lampepfl#15855 (attempt 2)