Fix substParam(s)
The `substParam` and `substParams` overlooked a case: If we substitute P := T and
we have a type variable A with origin P then A should map to T as well. Previously,
it didn't  which means that substParam(s) could break the link between a type
variable and the type parameter it stands for. This gives more freedoms to instantiate
type variables than allowed so it can mask type errors.