Give GadtConstraint a nicer toText
Instead of showing the underlying constraint info, which just looks bad,
takes up lots of space, and shows details that aren't relevant and often
includes leaked type constructors, just show the symbols and their full
bounds, nice and compact.  And do that as a part of toText instead of
having a side "debugBoundsDescription" alternative.