JavaParsers: Java interfaces should not have the Abstract flag
The Trait and Abstract flags are not supposed to be set together, and we don't
set Abstract when deserializing Java interfaces from classfiles so we shouldn't
do it when reading them from source code either.

While we're at it, this commit also drops Flags.Trait which is implied by Flags.JavaInterface.