Eliminate imports after cross version checks
... except for language imports. It looks this does not change the underlying issue
but it's good housekeeping anyway. The problem is that imports would not be transformed
in a phase like ExplicitOuter since Import is not by default handled in MegaPhase's
transform. As long as no-one checks this might not matter, but it's still weird. So
better to get rid of them before.