Keep `_` as a "don't know" in type patterns
E.g. in

    case x: Either[_, T] =>

treat the `_` as a "don't know" wildcard instead of turning the `Either` into a type lambda.
This applies under -language:experimental.namedTypeArgs and under -Ykind-projector:underscores
which both normally treat `_` in types as lambda placeholders.

For patterns, it was considered that the equivalence that `C[_] = C[t]` for some fresh type
variable `t` is more important.