Drop `toNoExplanation` implicit conversion
`toNoExplanation` was an old-style implicit conversion. It could not be made a
`Conversion` since it was defined over a call-by-name argument. It also hid
important implementation details since it delayed constructing the argument.
So knowing whether a string would be evaluated strictly or lazily depended on
knowing how type inference would work.

We replace the conversion by a combination of

 - an extension method `.toMessage` converting a lazy string explicitly, and
 - some overloading of core methods that take either a message or a lazy string.

We should over time get rid of overloads and convert most strings passed as messges
to proper messages. But that will take time. At least by now we see better where
the work is needed.