Fix Text wrapping
There was a problem with how some text wrapped.  I noticed it in the
printing of big tuple types and I started to write unit tests for those,
but there turned out to also be a checkfile that demonstrates the fix:
tests/neg/i14127.check:

    -  |  Int
    -  |, Int, Int)] was found for parameter x [...]
    -  |  Int
    -  |, Int, Int)]:
    +  |  Int, Int, Int)] was found for parameter [...]
    +  |  Int, Int, Int)]:

The fix is to make appendIndented, which is used by append, use Fluid
instead of switching to Vertical.  It seems to me that, pre-layout,
Vertical means "I want my elements to be vertical", i.e. wrap every one,
and it's layout that takes that into account.  Fluid instead means "wrap
where necessary, pack where possible".  Post-layout, instead, it looks
like Fluid means one-per-line, as it's consumed by print.  So the fix is
to switch appendIndented from Vertical to Fluid.

Following that I made some improvements like not needless boxing
non-splittable text, like Str, into Closed and merging non-closed Fluid
text during concat (`~`).

That fixed the extra wrapping, but it meant that the ", " separator
could be the text that is wrapped and indented, which didn't look nice.
So I changed Text.apply to close the separator to the previous element.
That encouraged lines ending with ", ", that is with a trailing space,
so I made print strip trailing spaces.

The change from Vertical to Fluid made the last parenthesis in
tests/neg/i9185.check not wrap, which is fine, but that pulled the
trailing "failed with" up to the previous line.  As that's not code, I
decided to move it down and a line away from the code.