Fix errors in explicit type annotations in inline match cases
Previusly, Unapply trees would have type bindings generated inside their
body and this was the only case handled in InlineReducer. However,
this happened for inferred type parameters, and Unapply with an
explicit binding inside a type annotation was not handled, leading to
a "cannot reduce match" error. This case is now handled and a related
comment was added as well.