Reduce implicit search space for BuildFrom avoiding divergence
BuildFrom (nee CanBuildFrom) is de-emphasised in 2.13 but remains
for use by generic extension methods that abstract over different
collection types.

As part of the 2.13 redesign, implicit instances for
`BuildFrom[SomeCollection]` where moved from the `SomeCollection`
companions into the `BuildFrom` companion.

This has an unfortunate side effect -- more implicits are in
scope for implicit searches, which can incur longer search time
and can also lead into divergent searches.

Implicit search prunes candidates based on the "depoly"-ed type,
which type params and value params with wildcards types in
the implicit signature. The nested implicit search for implicit
params occurs, which can guide inference of the type params.
Only then is the true type known and compared against the
expected type of the search.

It seems that this "depoly" transform could be smarter by
approximating type params to bounds. Perhaps this is harder
than it sounds when you have things like F-bounded or mutually
recursive type params.

Instead of trying to fix that general problem, I've instead
restated the bounds in the result type of relevant implicits
in BuildFrom. This would be redundant, were it not for way
that they are preseved through `depoly`.

The new test case now compiles without divergence and
the search considers only one candidate in detail.

Previously, compile failed with a divergent error in
`Ordering` implicits after about 80 steps.