Avoid interpolating to `Nothing` in a special case
If we have an selection `e.m` where e's type is a type variable T that does not
have a lower bound in the current constraint, avoid interpolating `T` to nothing.
Doing this would make the section ill-typed. By both interpolating `T` we leave
the possibility open to instantiate it to its upper bound in couldInstantiateTypeVar.