Sync reference even if there are conflicts.
Right now when there are conflicts between the language-reference-stable
branch and the main, the CI is not creating sync PR. This results in
growing gap between the stable reference and the main.

I've changed the CI behavior, so the PR is always created. The person
that is the author of last commit is assigned so they can fix potential
conflicts.