Avoid stackoverflow in ExplicitOuter
When transforming a class at ExplicitOuter we create outer accessors for it.
The newSymbol call to do this takes place at phase ExplicitOuter + 1, but
its arguments need to be evaluated at phase ExplicitOuter. This was not true
for the nestingLevel argument, which demanded the denotation of the class
at phase ExplicitOuter + 1, thus leading to the SO.

Interestingly, the same path is not taken if the class has all abstract members defined
or is declared abstract. It's only in the error case that I could reproduce the SO.