Take @targetName into account when resolving extension methods of value classes
Before target name we only matched on signatures. This was OK, since multiple extension
methods of the same class must be different, otherwise we will get a "have the same erasure"
error later at erasurePhase. But with @targetName that's now a legal situation that
needs to be resolve correctly. We do this by propagating the target name to the extension
method and verifiying that the target names of the original and extension methods match.

Fixes #16464