Trial balloon: Make default getters final
I want to see what the fallout is when we make default getters final, which
is necessary if we want to make them inlineable.

So far I have seen ~20 occurrences in the compiler codebase itself and a handful more
in the tests where an overriding function has exactly the same default argument as the
overridden function, so the repeated default argument was redundant.

I have seen no occurrences in this codebase where the overriding function had a
meaningful different default argument, except for tests that were specifically designed
to test that feature.