[Proof of Concept] Code generation via rewriting errors in macro annotations
This commit prototypes the proposal in
https://contributors.scala-lang.org/t/scala-3-macro-annotations-and-code-generation/6035
to use a macro annotation to do code generation. Since the -rewrite mechanism
isn't currently accessible to macros, we just print an error indicating what
part of the source code needs to be rewritten.

We implement support for a subset of `@data`: only the `withField` methods are
added (note that equals/hashCode/toString wouldn't need to be code-generated
since they are already defined in `AnyRef`, so a macro annotation can override
them as demonstrated in the existing `tests/run-macros/annot-mod-class-data`
test).

This is only one possible path forward. Alternatively, we could recommend using
scalafix for this usecase, and then the compiler itself wouldn't have to grow
code generation abilities, but libraries defining macro annotation might not
like to be tied to an external tool.