Add error handling strawman
There is so far no standard way in Scala to address a class of situations
that's often called "error handling". In these situations we have two possible
outcomes representing success and failure. We would like to process successful
results further and propagate failures outward. The aim is to find a "happy path" where
we deal with success computations directly, without having to special-treat failing
results at every step.

There are many different ways in use in Scala to express these situations.

 - Express success with normal results and failure with exceptions.
 - Use `Option`, where success is wrapped in `Some` and failure is `None`.
 - Use `Either`, where success is wrapped in `Right` and failure is wrapped in `Left`.
 - Use `Try`, where success is wrapped in `Success` and failure (which must be an exception) is wrapped in `Failure`.
 - Use nulls, where success is a non-null value and failure is `null`.
 - Use a special `IO` monad like `ZIO`, where errors are built in.

Exceptions propagate silently until they are handled in a `catch`. All other failure
modes require either ad-hoc code or monadic lifting into for comprehensions to propagate.
`nulls` can only be treated with ad-hoc code and they are not very popular in Scala so far.

Exceptions should only be used if the failure case is very unlikely (rule of thumb: in less than 0.1% of cases).
Even then they are often shunned since they are clunky to use and they currently undermine type safety (the latter
point would be addressed by the experimental `saferExceptions` extension). That said, exceptions have their uses.
For instance, it's better to abort code that fails in a locally unrecoverable way with an exception instead of
aborting the whole program with a panic or `System.exit`.

If we admit that not all error handling scenarios should be handled with exceptions, what
else should we use? So far the only choices are ad-hoc code or monadic lifting. Both of these
techniques suffer from the fact that a failure is propagated out of only a single construct
and that major and pointless acrobatics are required to move out failures further.

This commit sketches a systematic and uniform solution to this problem. It essentially provides a way
to establish "prompts" for result types and a way to return to a prompt with an error value using
a method called `_.?`. This is somewhat similar to Rust's `?` postfix operator, which seems
to work well. But there are is an important difference: Rust's ? returns from the next-enclosing
function or closure whereas our `?` method returns to an enclosing prompt, which can be further outside.
This greatly improves the expressiveness of `?`, at the price of a more complex execution semantics.

For instance, here is an implementation of a "traverse"-like operation, converting a `List[Option[T]]`
into a `Option[List[T]]`.
```scala
def traverse[T](xs: List[Option[T]]): Option[List[T]] =
  optional:
    xs.map(_.?)
```
Here, `optional` is the prompt for the `Option` type, and `.?` maps `Option[T]` to the underlying
type `T`, jumping to the enclosing `optional` prompt in the case a `None` is encountered. You can
think of it as a safe version of `Option`'s `get`. This function could not be written like this
in Rust since `_.?` is a closure so a `None` element would simply be mapped to itself.

Also unlike for Rust, Scala's technique is library-based. Similar prompt/return pairs can be defined
for many other types. This commit also defines a type `Result` which is similar to Rust's that
supports the pattern. `Result` will hopefully replace `Either`, which is in my opinion an abomination
for error handling. All these constructs can be defined in terms of a lower-level construct
where one can simply return a result to an enclosing boundary, without any distinction between
success and failure. The low-level construct is composed of a `boundary` prompt and a `break`
method that returns to it. It can also be used as a convenient replacement for non-local returns.

On the JVM, the break operation is in general implemented by throwing a special `Break` exception,
similar to how non-local returns were handled before. However, it is foreseen to implement an
optimization pass that replaces `Break` exceptions by gotos if they are caught in the same scope without
intervening functions. Dotty already has a `LabeledBlock` tree node that can be used
for this purpose. This means that local return to prompts are about as fast as in Rust whereas
non-local returns in Scala are a bit slower than local returns, whereas in Rust they are impossible.
We plan to do benchmarks to quantify the cost more precisely.