Use home-brewed futures for parallel pickling
Uses just one thread for the rest of pickling. One thread is sufficient since
there is not that much to do and we have time until the backend finishes.

We might want to partially revise that decision when we support pipelined
computation. In that case producing tasty early could be a win. But even in
that case we might want to fine-tune the number of worker threads instead
of relying on some executor. Adding more workers is easy in the new design.