Delay opaque alias checking until PostTyper
When we have a TypeBounds tree that also has an alias, we checked
that the alias is within bounds at Typer. This could provoke a
CyclicReference when processing opaque type aliases where
the type has a bound and also the alias refers back to the type.
We avoid the problem by performing the check at PostTyper, analogous
to other bounds checks.

Fixes #16642