Make sure annotations are typed in expression contexts
The previous logic tried but had holes. A local context was
established in `typedStats#localCtx` but that context was not
found in annotations that were typechecked in the completer
of the definition they decorated.

We are now more through and always allocate a localDummy as
new owner if the owner would otherwise we a class. This means
that closures in annotation arguments will have a term as owner,
and therefore will be allocated in distinct local scopes.

Fixes #15054