Change error position of additional params message
When there is an unexpected application, we had the function part
as position. This can be very misleading in the case of multi-line
curried functions, where we would see an arror pointing to the previous function
that complained that the function does not take parameters.

We now change that to report the _end_ of the previous function as error position.