Rename is... and info extension methods in Symbol
There are two separate problems with having `is` and `info` as extension methods in Symnbol.
For `is` methods: There's a "data race" error when compiling the compiler with itself. It
seems that both of the following conditions must hold for that error to come up:

 - the extension method must be overloaded
 - it must have a same-named real method in the target type that is
   reachable via an implicit conversion.

We need to minimize and dig deeper on this one.

For `info` the problem is different. We can't have both `info` and `info_=` as extension
methods since assignment desugaring does not work for extension methods. We cannot have `info`
alone either since then the conversion-based `info =` would not work anymore.

We should add a fix that allows assignment desugaring of extension methods to the language;
then this problem would go away.