Generate static inline accessors module
If a class `C` needs inline accessors that would be added top-level
or if the accessor is to a static member, we place it in a new
invisible module `C$inline$accessors`.

If the accessor location in the new scheme is not the same as the
previous location, we also generate the old accessor for backward binary
compatibility but do not use it.

Fixes #13215
Fixes #15413