Introduce Task abstraction
Tasks are Futures that have not started running yet. A task can be turned into a future by calling its
`run` method. Futures are now always eager, so `Future.start()` was dropped.

Tasks can be composed with `par` for parallel computations of pairs and `alt` for alternative
computations in case of errors.