Handle pickled forward references in pickled expressions
To compute forward references it was assumed that the owner of that
symbol can be found in the TASTy. This is not always the case with TASTy
expressions of pickled quotes. The owner might be outside the quote, in
this case the context already has the owner of the referenced symbol.
These are local symbols defined at the top-level of the TASTy.

Fixes #16843