Warn on use of inferred quote type variable bounds
This kind of inference is not reliable. We can only consider the bounds
from type constructor where the type variable is defined but any other
constraints are ignored. Therefore it is not possible to properly infer
the type bounds of the type variable.

The solution is simple, write the bounds explicitly and just check that
those bounds conform to the use site of the type variable.