Add `@binaryAPI`
The `@binaryAPI`  annotation will make any package private or protected term definition public in the generated bytecode. Definitions that override an `@binaryAPI` will also become public.

We cannot annotate `private[this]` definitions because we might have different definitions with the same name and signature in a subclass. These would clash if made public. Instead of using private/private[this], the user can write `private[C]` where `C` is the enclosing class.

This is useful in combination with inline definitions. If an inline definition refers to a `private`/`protected` definition marked as `@binaryAPI` it does not need to use an accessor. We still generate the accessors for binary compatibility but do not use them.